% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLM.R
\name{DLM}
\alias{DLM}
\title{Dynamic Linear Model Function on panel data}
\usage{
DLM(spec, data, model.name, contrb.date, group, date.var, start.date.m,
  end.date.m, start.date.c, end.date.c, is.output = F, is.graph = T,
  clv = 1, varlist, dum, dum_matrix, num_ng_coef, dep, actual, mean,
  CI.level = 0.95, iter = 100, level_type, var_coef, var_level, var_slope,
  per_sea, ord_sea, var_sea, optim.method = "L-BFGS-B", optim.lb = -Inf,
  optim.ub = Inf)
}
\arguments{
\item{spec}{a character vector for parallel computation; for example, rep("localhost",3) means open 3 local R sessions.}

\item{data}{a data.table class object as modelling dataset.}

\item{model.name}{a character for the name of your model, used in output file names.}

\item{contrb.date}{a character for the time period of your contribution calculation time window, used in output file names.}

\item{group}{a character for cross section variable name name in modelling dataset.}

\item{date.var}{a character for date variable name in modelling dataset.}

\item{start.date.m}{a character for the start date of modelling; format is YYYY-MM-DD.}

\item{end.date.m}{a character for the end date of modelling; format is YYYY-MM-DD.}

\item{start.date.c}{a character for the start date of post-modelling contribution calculation; format is YYYY-MM-DD.}

\item{is.output}{TRUE output decomp tables.}

\item{is.graph}{TRUE graph AVP chart.}

\item{clv}{a numeric for value calcualtion of dependent, such as NPV or CLV. Default value is 1.}

\item{varlist}{a character vector for the independent variable names, if no independent, then put c(). 
Always put variables required NEGATIVE coefficient at the beginning, then POSITIVE coefficient variables, and variables NOT required the sign of coefficient in the end.
The order would be used in a build-in coefficient sign check feature. If no sign check needed, then the order doesn't matter.}

\item{dum}{a character vector for variables in varlist which is not required sign(the last several variables from the above vector),if none then put c().}

\item{dum_matrix}{a numeric matrix for the variables in the dum vector, which specify the cross section that each variable would have impact on. If dum is c(), then put NA.
Row is cross section and column is variables in dum vector. Only put 1 or 0 to specify whether a variable has impact or not.}

\item{num_ng_coef}{a numeric for the number of variables required negative cofficient.}

\item{dep}{a character for dependent variable name.}

\item{actual}{a character for actual depdendent before indexed.}

\item{mean}{a character for weight used to index actual dependent.}

\item{CI.level}{a numeric for confidence interval level in output; range is from 0 to 1. For example, 0.95 means 95 percent CI level.}

\item{iter}{a numeric for the number of iteration of coefficient sign check. Default value is 100. If no sign check needed, please put 1.}

\item{level_type}{a numeric for the type of level component in DLM; 1 represents level type; 2 represents level+slope type.}

\item{var_coef}{a numeric for the variance of independent coefficient; 0 reprensents constant coef, non-zero value represents time-varying coef. For example, 1^2 means the standard deviation of each coef is 1 and time-varying;
"est" means estimated by model itself.}

\item{var_level}{a numeric for the variance of level component; 0 reprensents constant coef, non-zero value represents time-varying coef. For example, 1^2 means the standard deviation of each coef is 1 and time-varying;
"est" means estimated by model itself.}

\item{var_slope}{a numeric for the variance of slope component; 0 reprensents constant coef, non-zero value represents time-varying coef. For example, 1^2 means the standard deviation of each coef is 1 and time-varying;
"est" means estimated by model itself.}

\item{per_sea}{a numeric for the period of seasonality variable. For example, in weekly data, 52 means the period is 52 weeks.}

\item{ord_sea}{a numeric for the number of harmonics of seasonality component which is a Fourier series.}

\item{var_sea}{a numeric for the variance of seasonality component; 0 reprensents constant coef, non-zero value represents time-varying coef. For example, 1^2 means the standard deviation of each coef is 1 and time-varying;
"est" means estimated by model itself.}

\item{optim.method}{a charater for the optimization method used in MLE. Default is "L-BFGS-B". Any other method in stats::optim is available.}

\item{optim.lb}{a numeric for the lower bond of optimization. Defualt is -Inf.}

\item{optim.ub}{a numeric for the upper bond of optimization. Defualt is Inf.}

\item{end.date.m}{c character for the end date of post-modelling contribution calculation; format is YYYY-MM-DD.}
}
\value{
a list of modeling result. It contains the following components:
   stat: Diagnostic Statistics
   coef: a coefficient table if possible
}
\description{
A function to run DLM and post-modelling calculation on panel data for Marketing-Mix modelling
}

